+++ proxy/chunkreq.go
@@ -5,61 +5,17 @@ import (
 	"fmt"
 	"ghproxy/config"
 	"io"
-	"net"
 	"net/http"
 	"strconv"
-	"sync"
-	"time"
 
 	"github.com/gin-gonic/gin"
 )
-
-var BufferSize int = 32 * 1024 // 32KB
-
-var (
-	cclient    *http.Client
-	ctr        *http.Transport
-	BufferPool *sync.Pool
-)
-
-func InitReq(cfg *config.Config) {
-	initChunkedHTTPClient(cfg)
-	initGitHTTPClient(cfg)
-
-	// 初始化固定大小的缓存池
-	BufferPool = &sync.Pool{
-		New: func() interface{} {
-			return make([]byte, BufferSize)
-		},
-	}
-}
-
-func initChunkedHTTPClient(cfg *config.Config) {
-	ctr = &http.Transport{
-		MaxIdleConns:          100,
-		MaxConnsPerHost:       60,
-		IdleConnTimeout:       20 * time.Second,
-		TLSHandshakeTimeout:   10 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-		ResponseHeaderTimeout: 10 * time.Second,
-		DialContext: (&net.Dialer{
-			Timeout:   30 * time.Second,
-			KeepAlive: 30 * time.Second,
-		}).DialContext,
-	}
-	if cfg.Outbound.Enabled {
-		initTransport(cfg, ctr)
-	}
-	cclient = &http.Client{
-		Transport: ctr,
-	}
-}
 
 func ChunkedProxyRequest(c *gin.Context, u string, cfg *config.Config, mode string, runMode string) {
 	method := c.Request.Method
 
 	// 发送HEAD请求, 预获取Content-Length
-	headReq, err := http.NewRequest("HEAD", u, nil)
+	headReq, err := client.NewRequest("HEAD", u, nil)
 	if err != nil {
 		HandleError(c, fmt.Sprintf("Failed to create request: %v", err))
 		return
@@ -68,7 +24,7 @@ func ChunkedProxyRequest(c *gin.Context, u string, cfg *config.Config, mode stri
 	removeWSHeader(headReq) // 删除Conection Upgrade头, 避免与HTTP/2冲突(检查是否存在Upgrade头)
 	AuthPassThrough(c, cfg, headReq)
 
-	headResp, err := cclient.Do(headReq)
+	headResp, err := client.Do(headReq)
 	if err != nil {
 		HandleError(c, fmt.Sprintf("Failed to send request: %v", err))
 		return
@@ -92,13 +48,6 @@ func ChunkedProxyRequest(c *gin.Context, u string, cfg *config.Config, mode stri
 		}
 	}
 
-	/*
-		if err := HandleResponseSize(headResp, cfg, c); err != nil {
-			logWarning("%s %s %s %s %s Response-Size-Error: %v", c.ClientIP(), method, u, c.Request.Header.Get("User-Agent"), c.Request.Proto, err)
-			return
-		}
-	*/
-
 	body, err := readRequestBody(c)
 	if err != nil {
 		HandleError(c, err.Error())
@@ -107,30 +56,27 @@ func ChunkedProxyRequest(c *gin.Context, u string, cfg *config.Config, mode stri
 
 	bodyReader := bytes.NewBuffer(body)
 
-	// 创建请求
-	req, err := http.NewRequest(method, u, bodyReader)
+	req, err := client.NewRequest(method, u, bodyReader)
 	if err != nil {
 		HandleError(c, fmt.Sprintf("Failed to create request: %v", err))
 		return
 	}
-
 	setRequestHeaders(c, req)
 	removeWSHeader(req) // 删除Conection Upgrade头, 避免与HTTP/2冲突(检查是否存在Upgrade头)
 	AuthPassThrough(c, cfg, req)
 
-	resp, err := cclient.Do(req)
+	resp, err := client.Do(req)
 	if err != nil {
 		HandleError(c, fmt.Sprintf("Failed to send request: %v", err))
 		return
 	}
 	defer resp.Body.Close()
 
-	/*
-		if err := HandleResponseSize(resp, cfg, c); err != nil {
-			logWarning("%s %s %s %s %s Response-Size-Error: %v", c.ClientIP(), method, u, c.Request.Header.Get("User-Agent"), c.Request.Proto, err)
-			return
-		}
-	*/
+	// 错误处理(404)
+	if resp.StatusCode == 404 {
+		c.String(http.StatusNotFound, "File Not Found")
+		return
+	}
 
 	contentLength = resp.Header.Get("Content-Length")
 	if contentLength != "" {